Add operator set update fetching route to REST server (#21)
* fix: Fix operator set update ID type on MessageDatabase methods

* feat: Add operator set update route to REST server

* fix: Make REST API response structs fields public

+++ aggregator/message_database.go
@@ -95,7 +95,7 @@ func (md *MessageDatabase) StoreOperatorSetUpdate(operatorSetUpdateMessage regis
 	return md.Store("operatorSetUpdates", fmt.Sprintf("%d", operatorSetUpdateMessage.Id), operatorSetUpdateMessage)
 }
 
-func (md *MessageDatabase) FetchOperatorSetUpdate(id uint32, operatorSetUpdateMessage *registryrollup.OperatorSetUpdateMessage) error {
+func (md *MessageDatabase) FetchOperatorSetUpdate(id uint64, operatorSetUpdateMessage *registryrollup.OperatorSetUpdateMessage) error {
 	return md.Fetch("operatorSetUpdates", fmt.Sprintf("%d", id), operatorSetUpdateMessage)
 }
 
@@ -103,6 +103,6 @@ func (md *MessageDatabase) StoreOperatorSetUpdateAggregation(operatorSetUpdateMe
 	return md.Store("operatorSetUpdates", fmt.Sprintf("%d", operatorSetUpdateMessage.Id), aggregation)
 }
 
-func (md *MessageDatabase) FetchOperatorSetUpdateAggregation(id uint32, aggregation *types.MessageBlsAggregationServiceResponse) error {
+func (md *MessageDatabase) FetchOperatorSetUpdateAggregation(id uint64, aggregation *types.MessageBlsAggregationServiceResponse) error {
 	return md.Fetch("operatorSetUpdates", fmt.Sprintf("%d", id), aggregation)
 }
